+++ dart/example/bin/example.dart
@@ -40,14 +40,14 @@ Future<void> runApp() async {
   );
 
   Sentry.configureScope((scope) {
+    scope.setUser(SentryUser(
+      id: '800',
+      username: 'first-user',
+      email: 'first@user.lan',
+      // ipAddress: '127.0.0.1', sendDefaultPii feature is enabled
+      extras: <String, String>{'first-sign-in': '2020-01-01'},
+    ));
     scope
-      ..user = SentryUser(
-        id: '800',
-        username: 'first-user',
-        email: 'first@user.lan',
-        // ipAddress: '127.0.0.1', sendDefaultPii feature is enabled
-        extras: <String, String>{'first-sign-in': '2020-01-01'},
-      )
       // ..fingerprint = ['example-dart'], fingerprint forces events to group together
       ..transaction = '/example/app'
       ..level = SentryLevel.warning