+++ dart/test/mocks/mock_platform.dart
@@ -3,5 +3,15 @@ import 'package:sentry/src/platform/platform.dart';
 import 'no_such_method_provider.dart';
 
 class MockPlatform extends Platform with NoSuchMethodProvider {
-  const MockPlatform();
+  MockPlatform({String? os}) : operatingSystem = os ?? '';
+
+  factory MockPlatform.android() {
+    return MockPlatform(os: 'android');
+  }
+
+  @override
+  String operatingSystem;
+
+  @override
+  bool get isAndroid => (operatingSystem == 'android');
 }